refactor: GoalCardProps 정의 파일 변경 #32
